GenerationConfig, logits_process unittests
