+++ paddlenlp/utils/__init__.py
@@ -17,7 +17,7 @@ import contextlib
 import paddle
 
 from .batch_sampler import *
-from .env import CONFIG_NAME, LEGACY_CONFIG_NAME
+from .env import CONFIG_NAME, GENERATION_CONFIG_NAME, LEGACY_CONFIG_NAME
 from .import_utils import install_package, uninstall_package
 from .initializer import to
 from .serialization import load_torch
