+++ paddlenlp/utils/env.py
@@ -68,6 +68,7 @@ SUCCESS_STATUS = 0
 LEGACY_CONFIG_NAME = "model_config.json"
 CONFIG_NAME = "config.json"
 TOKENIZER_CONFIG_NAME = "tokenizer_config.json"
+GENERATION_CONFIG_NAME = "generation_config.json"
 
 
 LORA_CONFIG_NAME = "lora_config.json"
